requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
selenium==4.18.1
elasticsearch-dsl==8.9.0
python-dotenv==1.0.1
channels==4.0.0
django-cors-headers==4.3.1
transformers==4.37.2
spacy==3.7.2
pandas==2.1.4
numpy==1.26.3
huggingface-hub==0.20.3
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
pillow==10.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
boto3==1.34.45
requests==2.31.0





aiohttp==3.8.5
beautifulsoup4==4.12.2
tweepy==4.14.0
facebook-scraper==0.2.59
whatsapp-api-client-python==0.0.5


tweepy 
python-dotenv 
requests-html "lxml[html_clean]" facebook-scraper